Add Communication Profile Area template for POWERLINK Specification EPSG 301
This commit introduces a new template file for the Communication Profile Area (0x1000 - 0x1FFF) as defined by the POWERLINK Specification EPSG 301. The template includes references to various standards and includes default values according to the specifications. The initial version was created in 2010, with recent updates made to comments in June 2023.